Remove webpacker and related files
