+++ app/assets/stylesheets/application.bootstrap.scss
@@ -1,0 +1,17 @@
+@import "variables";
+@import "mixins";
+@import "fix_bootstrap_5_links";
+
+@import "bootstrap/scss/functions";
+
+@import "vendor/1st_load_framework";
+@import "vendor/animate";
+@import "vendor/simple-sidebar";
+
+@import "font_awesome5_webfont";
+
+// Custom bootstrap variables must be set or imported *before* bootstrap.
+@import "bootstrap/scss/bootstrap";
+
+@import "pages/*";
+@import "utilities/*";
