--- app/assets/stylesheets/application.scss
@@ -1,31 +1,0 @@
-/*
- * This is a manifest file that'll be compiled into application.css, which will include all the files
- * listed below.
- *
- * Any CSS and SCSS file within this directory, lib/assets/stylesheets, vendor/assets/stylesheets,
- * or any plugin's vendor/assets/stylesheets directory can be referenced here using a relative path.
- *
- * You're free to add application-wide styles to this file and they'll appear at the bottom of the
- * compiled file so the styles you add here take precedence over styles defined in any styles
- * defined in the other CSS/SCSS files in this directory. It is generally better to create a new
- * file per style scope.
- *
- */
-
-@import "variables";
-@import "mixins";
-@import "fix_bootstrap_5_links";
-
-@import "bootstrap/scss/functions";
-
-@import "vendor/1st_load_framework";
-@import "vendor/animate";
-@import "vendor/simple-sidebar";
-
-@import "font_awesome5_webfont";
-
-// Custom bootstrap variables must be set or imported *before* bootstrap.
-@import "bootstrap/scss/bootstrap";
-
-@import "pages/*";
-@import "utilities/*";
