--- app/javascript/packs/live_entry.js
@@ -1,1 +1,0 @@
-import "live_entry/index";
