+++ app/views/layouts/application.html.erb
@@ -8,7 +8,6 @@
   <% if @current_user %>
     <script id="current_user" type="text/json"><%= raw(::Api::V1::UserSerializer.new(@current_user).to_json) %></script>
   <% end %>
-  <%= javascript_pack_tag "application" %>
   <%= render "layouts/google_maps_api_js" %>
   <%= csrf_meta_tags %>
   <%= render "layouts/cloudflare_analytics" if Rails.env.production? %>
