--- config/webpack/development.js
@@ -1,3 +1,0 @@
-const environment = require('./environment')
-
-module.exports = environment.toWebpackConfig()
