+++ package.json
@@ -1,14 +1,12 @@
 {
   "dependencies": {
     "@hotwired/stimulus": "^3.0.1",
-    "@hotwired/stimulus-webpack-helpers": "^1.0.1",
     "@hotwired/turbo-rails": "^7.0.1",
     "@popperjs/core": "^2.11.7",
     "@rails/actioncable": "^7.0.2-2",
     "@rails/activestorage": "^6.0.5",
     "@rails/request.js": "^0.0.8",
     "@rails/ujs": "^7.0.4-3",
-    "@rails/webpacker": "^5.4.3",
     "@sentry/browser": "^7.46.0",
     "bootstrap": "^5.2.3",
     "bootstrap-notify": "^3.1.3",
@@ -26,13 +24,13 @@
     "underscore": "^1.13.6"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "webpack-dev-server": "^4.5.0"
+    "@babel/plugin-proposal-class-properties": "^7.7.4"
   },
   "engines": {
     "node": "16.x"
   },
   "scripts": {
-    "build": "esbuild app/javascript/*.* --bundle --sourcemap --outdir=app/assets/builds --public-path=assets"
+    "build": "esbuild app/javascript/*.* --bundle --sourcemap --outdir=app/assets/builds --public-path=assets",
+    "build:css": "sass ./app/assets/stylesheets/application.bootstrap.scss:./app/assets/builds/application.css --no-source-map --load-path=node_modules"
   }
 }
